+++ docs/src/stories/HOCReadme.js
@@ -1,18 +1,18 @@
-/* eslint-disable import/no-webpack-loader-syntax */
+/* eslint-disable  */
 import React from 'react'
 import marked from 'marked'
 //
-import Readme from '!raw!../../../README.md'
+import HOCReadme from '!raw!../../../src/hoc/README.md'
 import 'github-markdown-css/github-markdown.css'
 import './utils/prism.js'
 
-export default class Story extends React.Component {
+export default class HOCStory extends React.Component {
   render () {
     return (
       <div style={{ padding: '10px' }}>
         <span
           className='markdown-body'
-          dangerouslySetInnerHTML={{ __html: marked(Readme) }}
+          dangerouslySetInnerHTML={{ __html: marked(HOCReadme) }}
         />
       </div>
     )