+++ src/lifecycle.js
@@ -64,7 +64,7 @@ export default Base =>
       if (freezeWhenExpanded) {
         // if any rows are expanded, freeze the existing data and sorting
         const keys = Object.keys(newResolvedState.expanded)
-        for (var i = 0; i < keys.length; i++) {
+        for (let i = 0; i < keys.length; i += 1) {
           if (newResolvedState.expanded[keys[i]]) {
             newResolvedState.frozen = true
             break
@@ -99,21 +99,36 @@ export default Base =>
         Object.assign(newResolvedState, this.getSortedData(newResolvedState))
       }
 
+      // Set page to 0 if filters change
+      if (oldState.filtered !== newResolvedState.filtered) {
+        newResolvedState.page = 0
+      }
+
       // Calculate pageSize all the time
       if (newResolvedState.sortedData) {
         newResolvedState.pages = newResolvedState.manual
           ? newResolvedState.pages
           : Math.ceil(
-            newResolvedState.sortedData.length / newResolvedState.pageSize
+            newResolvedState.sortedData.length / newResolvedState.pageSize,
           )
         newResolvedState.page = Math.max(
           newResolvedState.page >= newResolvedState.pages
             ? newResolvedState.pages - 1
             : newResolvedState.page,
-          0
+          0,
         )
       }
 
-      return this.setState(newResolvedState, cb)
+      return this.setState(newResolvedState, () => {
+        if (cb) { cb() }
+        if (
+          oldState.page !== newResolvedState.page ||
+          oldState.pageSize !== newResolvedState.pageSize ||
+          oldState.sorted !== newResolvedState.sorted ||
+          oldState.filtered !== newResolvedState.filtered
+        ) {
+          this.fireFetchData()
+        }
+      })
     }
   }