+++ src/utils.js
@@ -19,6 +19,7 @@ export default {
   compactObject,
   isSortingDesc,
   normalizeComponent,
+  asPx,
 }
 
 function get (obj, path, def) {
@@ -29,7 +30,9 @@ function get (obj, path, def) {
   let val
   try {
     val = pathObj.reduce((current, pathPart) => current[pathPart], obj)
-  } catch (e) {}
+  } catch (e) {
+    // continue regardless of error
+  }
   return typeof val !== 'undefined' ? val : def
 }
 
@@ -58,7 +61,7 @@ function last (arr) {
 
 function range (n) {
   const arr = []
-  for (let i = 0; i < n; i++) {
+  for (let i = 0; i < n; i += 1) {
     arr.push(n)
   }
   return arr
@@ -66,7 +69,7 @@ function range (n) {
 
 function orderBy (arr, funcs, dirs, indexKey) {
   return arr.sort((rowA, rowB) => {
-    for (let i = 0; i < funcs.length; i++) {
+    for (let i = 0; i < funcs.length; i += 1) {
       const comp = funcs[i]
       const desc = dirs[i] === false || dirs[i] === 'desc'
       const sortInt = comp(rowA, rowB)
@@ -82,8 +85,8 @@ function orderBy (arr, funcs, dirs, indexKey) {
 }
 
 function remove (a, b) {
-  return a.filter(function (o, i) {
-    var r = b(o)
+  return a.filter((o, i) => {
+    const r = b(o)
     if (r) {
       a.splice(i, 1)
       return true
@@ -100,7 +103,7 @@ function clone (a) {
           return value.toString()
         }
         return value
-      })
+      }),
     )
   } catch (e) {
     return a
@@ -108,7 +111,7 @@ function clone (a) {
 }
 
 function getFirstDefined (...args) {
-  for (var i = 0; i < args.length; i++) {
+  for (let i = 0; i < args.length; i += 1) {
     if (typeof args[i] !== 'undefined') {
       return args[i]
     }
@@ -116,19 +119,20 @@ function getFirstDefined (...args) {
 }
 
 function sum (arr) {
-  return arr.reduce((a, b) => {
-    return a + b
-  }, 0)
+  return arr.reduce((a, b) => (
+    a + b
+  ), 0)
 }
 
 function makeTemplateComponent (compClass, displayName) {
   if (!displayName) {
     throw new Error('No displayName found for template component:', compClass)
   }
-  const cmp = ({ children, className, ...rest }) =>
+  const cmp = ({ children, className, ...rest }) => (
     <div className={classnames(compClass, className)} {...rest}>
       {children}
     </div>
+  )
   cmp.displayName = displayName
   return cmp
 }
@@ -142,6 +146,11 @@ function groupBy (xs, key) {
   }, {})
 }
 
+function asPx (value) {
+  value = Number(value)
+  return Number.isNaN(value) ? null : `${value}px`
+}
+
 function isArray (a) {
   return Array.isArray(a)
 }
@@ -153,8 +162,8 @@ function isArray (a) {
 function makePathArray (obj) {
   return flattenDeep(obj)
     .join('.')
-    .replace('[', '.')
-    .replace(']', '')
+    .replace(/\[/g, '.')
+    .replace(/\]/g, '')
     .split('.')
 }
 
@@ -162,7 +171,7 @@ function flattenDeep (arr, newArr = []) {
   if (!isArray(arr)) {
     newArr.push(arr)
   } else {
-    for (var i = 0; i < arr.length; i++) {
+    for (let i = 0; i < arr.length; i += 1) {
       flattenDeep(arr[i], newArr)
     }
   }
@@ -173,20 +182,23 @@ function splitProps ({ className, style, ...rest }) {
   return {
     className,
     style,
-    rest,
+    rest: rest || {},
   }
 }
 
 function compactObject (obj) {
   const newObj = {}
-  for (var key in obj) {
-    if (
-      obj.hasOwnProperty(key) &&
-      obj[key] !== undefined &&
-      typeof obj[key] !== 'undefined'
-    ) {
-      newObj[key] = obj[key]
-    }
+  if (obj) {
+    Object.keys(obj).map(key => {
+      if (
+        Object.prototype.hasOwnProperty.call(obj, key) &&
+        obj[key] !== undefined &&
+        typeof obj[key] !== 'undefined'
+      ) {
+        newObj[key] = obj[key]
+      }
+      return true
+    })
   }
   return newObj
 }